+++ bigchaindb/core.py
@@ -7,7 +7,7 @@ import bigchaindb
 from bigchaindb import config_utils
 from bigchaindb import exceptions
 from bigchaindb import util
-from bigchaindb.crypto import core
+from bigchaindb.crypto import asymmetric
 from bigchaindb.monitor import Monitor
 
 monitor = Monitor()
@@ -96,7 +96,7 @@ class Bigchain(object):
 
         signature = data.pop('signature')
         public_key_base58 = signed_transaction['transaction']['current_owner']
-        public_key = core.PublicKey(public_key_base58)
+        public_key = asymmetric.PublicKey(public_key_base58)
         return public_key.verify(util.serialize(data), signature)
 
     @monitor.timer('write_transaction', rate=bigchaindb.config['statsd']['rate'])
@@ -329,8 +329,8 @@ class Bigchain(object):
 
         # Calculate the hash of the new block
         block_data = util.serialize(block)
-        block_hash = core.hash_data(block_data)
-        block_signature = core.PrivateKey(self.me_private).sign(block_data)
+        block_hash = asymmetric.hash_data(block_data)
+        block_signature = asymmetric.PrivateKey(self.me_private).sign(block_data)
 
         block = {
             'id': block_hash,
@@ -355,7 +355,7 @@ class Bigchain(object):
         """
 
         # 1. Check if current hash is correct
-        calculated_hash = core.hash_data(util.serialize(block['block']))
+        calculated_hash = asymmetric.hash_data(util.serialize(block['block']))
         if calculated_hash != block['id']:
             raise exceptions.InvalidHash()
 
@@ -450,7 +450,7 @@ class Bigchain(object):
         }
 
         vote_data = util.serialize(vote)
-        signature = core.PrivateKey(self.me_private).sign(vote_data)
+        signature = asymmetric.PrivateKey(self.me_private).sign(vote_data)
 
         vote_signed = {
             'node_pubkey': self.me,
