+++ tests/web/test_basic_views.py
@@ -3,7 +3,7 @@ import json
 import pytest
 
 from bigchaindb import util
-from bigchaindb.crypto import core
+from bigchaindb.crypto import asymmetric
 
 
 TX_ENDPOINT = '/api/v1/transactions/'
@@ -18,7 +18,7 @@ def test_get_transaction_endpoint(b, client, user_public_key):
 
 
 def test_post_create_transaction_endpoint(b, client):
-    keypair = core.generate_key_pair()
+    keypair = asymmetric.generate_key_pair()
 
     tx = util.create_and_sign_tx(keypair[0], keypair[1], keypair[1], None, 'CREATE')
 
@@ -28,8 +28,8 @@ def test_post_create_transaction_endpoint(b, client):
 
 
 def test_post_transfer_transaction_endpoint(b, client):
-    from_keypair = core.generate_key_pair()
-    to_keypair = core.generate_key_pair()
+    from_keypair = asymmetric.generate_key_pair()
+    to_keypair = asymmetric.generate_key_pair()
 
     tx = util.create_and_sign_tx(from_keypair[0], from_keypair[1], from_keypair[1], None, 'CREATE')
     res = client.post(TX_ENDPOINT, data=json.dumps(tx))
